# docs/swagger.yaml
definitions:
  requester.EditShortUrl:
    properties:
      ab_percent:
        type: integer
      source_url_a:
        type: string
      source_url_b:
        type: string
      tracker_id:
        type: string
      who_click:
        type: boolean
    required:
    - tracker_id
    type: object
  requester.ShortenUrl:
    properties:
      source_url:
        type: string
    required:
    - source_url
    type: object
  responser.CheckUrlSafe:
    properties:
      result:
        type: boolean
      type:
        type: string
    type: object
  responser.Response:
    properties:
      data: {}
      message:
        type: string
      result_code:
        type: integer
      time_stamp:
        type: string
    type: object
host: thrurl-center.jmh-su.com
info:
  contact:
    email: junminhong1110@gmail.com
    name: junmin.hong
    url: https://github.com/junminhong
  description: 一個簡單易用且強大的縮網址服務
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Thrurl API
  version: "1.0"
paths:
  /api/v1/short-url:
    get:
      consumes:
      - application/json
      parameters:
      - default: Bearer <請在這邊輸入Atomic Token>
        description: Atomic Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: tracker-id
        in: query
        name: tracker-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: 請依照API文件進行請求
          schema:
            type: string
        "1001":
          description: 短連結生成失敗
          schema:
            type: string
        "1002":
          description: 短連結生成成功
          schema:
            $ref: '#/definitions/responser.Response'
        "1003":
          description: 無效連結
          schema:
            type: string
      summary: 取得短連結的訊息
      tags:
      - short-url
    post:
      consumes:
      - application/json
      parameters:
      - default: Bearer
        description: Atomic Token
        in: header
        name: Authorization
        type: string
      - description: 請求資料
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requester.ShortenUrl'
      produces:
      - application/json
      responses:
        "1000":
          description: 請依照API文件進行請求
          schema:
            type: string
        "1001":
          description: 短連結生成失敗
          schema:
            type: string
        "1002":
          description: 短連結生成成功
          schema:
            $ref: '#/definitions/responser.Response'
        "1003":
          description: 無效連結
          schema:
            type: string
      summary: 縮短網址
      tags:
      - short-url
    put:
      consumes:
      - application/json
      parameters:
      - default: Bearer <請在這邊輸入Atomic Token>
        description: Atomic Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 請求資料
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requester.EditShortUrl'
      produces:
      - application/json
      responses:
        "1000":
          description: 請依照API文件進行請求
          schema:
            type: string
        "1001":
          description: 短連結保存失敗
          schema:
            type: string
        "1002":
          description: 短連結保存成功
          schema:
            $ref: '#/definitions/responser.Response'
        "1003":
          description: 無效連結
          schema:
            type: string
        "1004":
          description: 你沒有權限發起該請求
          schema:
            type: string
        "1005":
          description: 找不到該連結訊息
          schema:
            type: string
      summary: 編輯短連結
      tags:
      - short-url
  /api/v1/short-url/click-info:
    get:
      consumes:
      - application/json
      parameters:
      - default: Bearer <請在這邊輸入Atomic Token>
        description: Atomic Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: tracker-id
        in: query
        name: tracker-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1004":
          description: 你沒有權限發起該請求
          schema:
            type: string
        "1005":
          description: 找不到該短連結的點擊成效
          schema:
            type: string
        "1011":
          description: 成功取得點擊成效
          schema:
            $ref: '#/definitions/responser.Response'
      summary: 取得短連結的點擊成效
      tags:
      - short-url
  /api/v1/short-url/list:
    get:
      consumes:
      - application/json
      parameters:
      - default: Bearer <請在這邊輸入Atomic Token>
        description: Atomic Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1004":
          description: 你沒有權限發起該請求
          schema:
            type: string
        "1006":
          description: 取得短連結列表
          schema:
            $ref: '#/definitions/responser.Response'
      summary: 取得短連結的列表
      tags:
      - short-url
  /api/v1/short-url/redirect:
    get:
      consumes:
      - application/json
      parameters:
      - description: tracker-id
        in: query
        name: tracker-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1010":
          description: 成功取得原始連結
          schema:
            $ref: '#/definitions/responser.Response'
      summary: 取得轉址要去的目標
      tags:
      - short-url
  /api/v1/url/check-safe:
    get:
      consumes:
      - application/json
      parameters:
      - description: source-url
        in: query
        name: source-url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1005":
          description: 不是有效連結
          schema:
            type: string
        "1007":
          description: 檢查完成
          schema:
            $ref: '#/definitions/responser.CheckUrlSafe'
      summary: 檢查網址安全
      tags:
      - url
  /api/v1/url/record:
    get:
      consumes:
      - application/json
      parameters:
      - description: tracker-id
        in: query
        name: tracker-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: 請依照API文件發起請求
          schema:
            type: string
        "1008":
          description: 記錄失敗
          schema:
            type: string
        "1009":
          description: 記錄完成
          schema:
            $ref: '#/definitions/responser.CheckUrlSafe'
      summary: 記錄點擊成效
      tags:
      - url
swagger: "2.0"